fix: create consistent navigation for public pages
- Created PublicNav component for shared navigation
- Both landing (/) and about (/about) pages now use the same nav
- Added Home link to nav and proper active state highlighting
- Consistent styling and behavior across public pages
- Removed unused auth imports from pages
